Restructured StringEquationParser -> Equation, basically swapping inner/outer class
